- Creating branch issue30 - Adding check for available port when launching OPDS publisher
Update issue 30
First step: added the feature of trying for available ports until it works.
I also added a "get_url" function that should return the URL where the catalog is availble, but I'm not sure how to display it and whether it will work because of the multi-process stuff.