- Creating branch issue37 - Adding fix to avoid repeated books (with the same ISBN) in the Alibris guesser
Update issue 37
The thing here is that in Alibris a book represents an actual book, so if the same book is sold in different libraries or in different usage conditions, you get more than one record.
I added a check so that books with the same ISBN are not repeated.
You will still see repeated books, but with different ISBNs (maybe with different publishers or publication dates).
Maybe we can display this additional information in the new guesser UI (issue 34).

Can you review this please?