- Changing the way of checking the available address:port for the OPDS catalog, switching to a simple socket - Adding a Queue to share information between the two processes - Adding a notification for the user with the URL were the catalog is published
Update issue 30
In the end it was easier to use a good old socket to check for host/port availability than hacking bottle.
The issue is fixed now. Can you review it?

New issue
Summary: Add posibility to stop the OPDS catalog server
Labels: Type-Enhancement Priority-Medium
We should add the posibility of stopping the OPDS catalog server without the need closing Aranduka.
One possibility would be to change the menu's action from "Publish catalog" to "Stop catalog server" or something like that.
But that raises the issue of how to implement that within the plugin structure.