- Passing additional download links to ManyBooks.net store plugin template - Adding error notice when ManyBooks.net page cannot be loaded
Update issue 46
I moved forward a little with this. I'm still having random problems to connect to ManyBooks.net, so I added a message for the user.
Now that ManyBooks provides a long list of available formats, we should probably change the way the links are displayed.
Maybe we could use a <select> with a Download button. I need to see how to do this on webkit.