- Creating branch issue55 - Changing name of menu option from "Open %s" to "Open book" when there's only one file - Fixing/cleaning the context menu
Update issue 55
Status: Started
I changed the title of the option to "Open book" when there's only one file and also made some fixes to the menu which wasn't working properly.
We still need to decide the best way to display the list of available files when there is more than one file in the same format.