Adding the possibility to remove a book from the database but not from the filesystem.
Update issue 22
I've fixed this. I replaced the QMessageBox.question with a new custom dialog which asks you if you want
to delete the book and lets you mark if you want the files to be deleted as well (checked by default).
I also made the necessary changes on Book and File models.